streamlit_app/requirements.txt
streamlit==1.29.0
tf-keras==2.16.0
numpy==1.24.3
Pillow==10.0.0
requests==2.31.0
